Multiple student entry
Teachers can now add students in bulk from a CSV file. Aeries format
needs to be followed (names in column C, last name first with comma
separating). Button added to student list for adding all students a
teacher has, as well as a button in the classroom for adding a bulk of
new students to a class.