Teachers can link parents, send e-mails
Teachers can put in the e-mail associated with a parent account to link
it with a specific teacher. If the e-mail is not associated with a
specific account, then an e-mail will be sent to that account requesting
that a ClassMinder account is made. PHPMailer was added to the build.