Client feedback for deletion, record behaviors
Users now get a confirmation window when deleting students and classes.
The record behaviors page was changed to include an alert message when a
user doesn't select students, and the entire student button is now
clickable to select a student. The student name column in the provided
roster file was split into first name and last name columns, and the
excel input code was edited to accommodate this.